requirements.txt
gensim
requests
python-telegram-bot
